restructure event reducer defs to be more concise
