+++ src/shared/gameEvents/allEventDefs.ts
@@ -1,0 +1,4 @@
+import create from './eventDefs/create';
+export default {
+  create,
+};
