+++ src/shared/gameEvents/eventDefs/create.ts
@@ -1,0 +1,11 @@
+import {EventDef} from '../types/EventDef';
+
+interface CreateJson {}
+
+const create: EventDef<CreateJson> = {
+  reducer(state, event) {
+    return state;
+  },
+};
+
+export default create;
