+++ src/shared/gameEvents/types/EventDef.ts
@@ -1,0 +1,5 @@
+import {GameState} from './GameState';
+
+export interface EventDef<P extends {}> {
+  reducer(state: GameState, params: P): GameState;
+}
