+++ src/shared/gameEvents/types/ExtractParamsType.ts
@@ -1,0 +1,3 @@
+import {EventDef} from './EventDef';
+
+export type ExtractParamsType<T> = T extends EventDef<infer R> ? R : unknown;
