+++ src/shared/gameEvents/types/GameEventType.ts
@@ -1,0 +1,3 @@
+import allEventDefs from '../allEventDefs';
+
+export type EventType = keyof typeof allEventDefs;
