+++ src/shared/gameEvents/types/GameState.ts
@@ -1,0 +1,6 @@
+import {GameJson} from '../../types';
+
+export interface GameState {
+  loaded: boolean;
+  game: null | GameJson;
+}
