--- src/shared/gameEvents/types/index.ts
@@ -1,3 +1,0 @@
-export * from './GameEvent';
-export * from './userServerPing';
-export * from './initializeGameEvent';
