--- src/shared/gameEvents/types/initializeGameEvent.ts
@@ -1,4 +1,0 @@
-import {GameEventType} from './GameEvent';
-import {makeGameEventTypeGuard} from './makeGameEventTypeGuard';
-
-export const isInitializeGameEvent = makeGameEventTypeGuard(GameEventType.initializeGame);
